Version 0.8 - première version utilisable

+++ BeCG/data/scripts.js
@@ -21,6 +21,7 @@
 
 var suiteOK    = false;
 var attenteOK  = false;
+var connectOK  = false;
 var scale_baOK = false
 var scale_ba   = 0.0;
 var scale_bfOK = false
@@ -32,6 +33,17 @@ function page_onload() {
   get_settings();
   // Affichage dynamique des valeurs de la balance
   setTimeout(function() { XMLHttpRequest_get("getvalues") }, 500);
+
+  //setTimeout(function() { alert(window.innerWidth + " - " + document.body.clientWidth) }, 1000);
+
+}
+
+function netconfig_onload() {
+  // Mise à jour des infos de la balance
+  get_settings();
+  // Retrouve la liste des réseaux vsibles
+  //XMLHttpRequest_get("getnetworks");
+  get_networks();
 }
 
 function get_settings() {
@@ -41,6 +53,186 @@ function get_settings() {
   XMLHttpRequest_get("getwifi");
   // Recupère les paramètres de la balance
   XMLHttpRequest_get("getsettings");
+}
+
+async function get_networks() {
+  // Affiche le fond gris
+  var dlgMask = document.getElementById("dlgMask0");
+  dlgMask.classList.remove("noshow");
+  var divWait = document.getElementById("attente0");
+  // Affiche l'annimation d'attente
+  divWait.classList.remove("noshow");
+  // envoi la requette au serveur et attend la réponse
+  attenteOK = false;
+  XMLHttpRequest_get("getnetworks");
+  while (!attenteOK) {
+    await sleep(100);
+  }
+  attenteOK = false;
+  // Masque l'animation d'attente
+  divWait.classList.add("noshow");
+  // Masque le fond gris
+  dlgMask.classList.add("noshow");
+}
+
+function setNetworkList(xml) {
+  // Fonction appelée en retour de XMLHttpRequest_get("getnetworks")
+  var htmlNetworkTable = "";
+  var tmpSSID      = "";
+  var tmpChannel   = 0;
+  var tmpRSSI      = "";
+  var tmpCrypt     = "";
+  var tmpImgSignal = "";
+  
+  var divNetworkTable = document.getElementById("networkTable");
+  
+  htmlNetworkTable  = "<table class=\"netTables\">\n";
+  htmlNetworkTable += "  <thead>\n";
+  htmlNetworkTable += "    <tr>\n";
+  htmlNetworkTable += "      <th>&nbsp;</th>\n";
+  htmlNetworkTable += "      <th>SSID</th>\n";
+  htmlNetworkTable += "      <th>Canal</th>\n";
+  htmlNetworkTable += "      <th>Sécurité</th>\n";
+  htmlNetworkTable += "    </tr>\n";
+  htmlNetworkTable += "  </thead>\n";
+  htmlNetworkTable += "  <tbody>\n";
+
+  var netListe = xml.getElementsByTagName("network")
+  if (netListe.length > 0) {
+    for(var i = 0; i< netListe.length; i++){
+      // Données du réseau
+      tmpSSID    = netListe[i].getElementsByTagName("SSID")[0].childNodes[0].nodeValue;
+      tmpChannel = netListe[i].getElementsByTagName("channel")[0].childNodes[0].nodeValue;
+      tmpRSSI    = netListe[i].getElementsByTagName("RSSI")[0].childNodes[0].nodeValue;
+      tmpCrypt   = netListe[i].getElementsByTagName("encryption")[0].childNodes[0].nodeValue;
+      // Choix du pictogramme en fonction de la qualité du signal
+      // RSSI (dBm) Interprétation
+      // -30 dBm    Extraordinaire (êtes vous assis sur la borne? ^^)
+      // -67 dBm    Très bon signal
+      // -70 dBm    Très acceptable
+      // -80 dBm   	as terrible du tout
+      // -90 dBm    Inutilisable
+      if (tmpRSSI <= -90) {
+        tmpImgSignal = "signal0.svg";
+      } else if ((tmpRSSI > -90) && (tmpRSSI <= -80)) {
+        tmpImgSignal = "signal1.svg";
+      } else if ((tmpRSSI > -80) && (tmpRSSI <= -70)) {
+        tmpImgSignal = "signal2.svg";
+      } else if ((tmpRSSI > -70) && (tmpRSSI <= -67)) {
+        tmpImgSignal = "signal3.svg";
+      } else if (tmpRSSI > -67) {
+        tmpImgSignal = "signal4.svg";
+      }
+      htmlNetworkTable += "    <tr class=\"trlink\" onclick=\"wifi_connect('" + tmpSSID + "' , '" + tmpChannel + "')\">\n";
+      htmlNetworkTable += "      <td class=\"centreVertical\"><img src=\"" + tmpImgSignal + "\" title=\"RSSI = " + tmpRSSI + "\" /></td>\n";
+      htmlNetworkTable += "      <td>" + tmpSSID + "</td>\n";
+      htmlNetworkTable += "      <td>" + tmpChannel + "</td>\n";
+      htmlNetworkTable += "      <td>" + tmpCrypt + "</td>\n";
+      htmlNetworkTable += "    </tr>\n";
+    }
+  } else {
+    htmlNetworkTable += "    <tr>\n";
+    htmlNetworkTable += "      <td colspan=\"4\">Aucun réseau disponible</td>";
+    htmlNetworkTable += "    </tr>\n";
+  }
+  htmlNetworkTable += "  </tbody>\n";
+  htmlNetworkTable += "</table>\n";
+  
+  divNetworkTable.innerHTML = htmlNetworkTable;
+  
+}
+
+async function wifi_connect(SSID, channel) {
+  // TODO saisie du mot de passe et confirmation 
+  // TODO voir pourquoi on ne reçois pas de réponse XML
+  // ? le serveur perd les infos du client lors de la déconnexion ?
+  var pwd = "C'est1secret";
+
+  var ssid_input = document.getElementById("ssid_input");
+  var pwd_input  = document.getElementById("pwd_input");
+  var divWait    = document.getElementById("attente0");
+  var btnConnect = document.getElementById("btnConnect");
+  var btnAnnuler = document.getElementById("btnAnnuler");
+
+  ssid_input.value = SSID + " (ch." + channel + ")";
+  
+  connectOK = true;
+  suiteOK = false;
+
+  // Affiche la boite de dialogue de saisie du mot de passe réseau
+  afficheDialog('dlgConnect');
+  // Donne le focus au champ de saisie du mot de passe
+  pwd_input.focus();
+
+  // Attente click sur bouton Connexion
+  while (!suiteOK) {
+    await sleep(100);
+  }
+  suiteOK = false;
+  
+  if (connectOK) {
+    pwd = pwd_input.value;
+    // Désactive les boutons de la boite de dialogue
+    btnConnect.disabled = true;
+    btnAnnuler.disabled = true;
+    // Affiche l'annimation d'attente
+    divWait.classList.remove("noshow");
+    // Lance la connexion de la balance
+    //alert("XMLHttpRequest_post_wificonnect(" + SSID + ", " + pwd + ", " + channel + ");");
+    XMLHttpRequest_post_wificonnect(SSID, pwd, channel);
+    // On fermera la boite de dialogue et on masquera
+    // l'animation d'attente dans 10 secondes
+    setTimeout(function() { divWait.classList.add("noshow"); }, 10000);
+    setTimeout(function() { closeDialog('dlgConnect'); location.reload(); }, 10000);
+  } else {
+    alert("Connexion annulée.");
+    closeDialog('dlgConnect');
+  }
+
+}
+
+function connect_clique() {
+  connectOK = true;
+  suiteOK   = true;
+}
+
+function connect_cancel() {
+  connectOK = false;
+  suiteOK   = true;
+}
+
+function connect_keyup(e) {
+  if (e.keyCode === 13) {
+    document.getElementById("btnConnect").click();
+  } else if (e.keyCode === 27) {
+    document.getElementById("btnAnnuler").click();
+  }
+}
+
+function deconnect_clique() {
+  if (confirm("Voulez-vous vraiment vous déconnecter du réseau en cours\net effacer les paramètres WiFi client dans la balance ?")) {
+    XMLHttpRequest_get("deconnexion");
+  }
+}
+
+function XMLHttpRequest_post_wificonnect(SSID, pwd, channel) {
+  var xhttp = new XMLHttpRequest();
+
+  xhttp.onreadystatechange = function() {
+    if (xhttp.readyState == 4) {
+      if ((xhttp.status == 200) || (xhttp.status == 0)) {
+        XMLHttpResult("wificonnect", xhttp.responseXML, xhttp.responseText);
+      } else {
+        alert("XMLHttpRequest_post_wificonnect() : Erreur " + xhttp.status);
+      }
+    }
+  };
+
+  var ssid_encode = encodeURIComponent(SSID);
+  var pwd_encode  = encodeURIComponent(pwd);
+  xhttp.open("POST", "wificonnect", true);
+  xhttp.setRequestHeader("Content-Type", "application/x-www-form-urlencoded");
+  xhttp.send("ssid=" + ssid_encode + "&pwd=" + pwd_encode + "&channel=" + channel);
 }
 
 function XMLHttpRequest_post_settings() {
@@ -56,14 +248,12 @@ function XMLHttpRequest_post_settings() {
       } else {
         alert("XMLHttpRequest_post_settings() : Erreur " + xhttp.status);
       }
-      //alert("OK"); // C'est bon \o/
     }
   };
 
   xhttp.open("POST", "setsettings", true);
   xhttp.setRequestHeader("Content-Type", "application/x-www-form-urlencoded");
   xhttp.send("paf_ba=" + paf_ba + "&entraxe=" + entraxe + "&tare=" + tare);
-
 }
 
 function XMLHttpRequest_get(requette) {
@@ -112,6 +302,17 @@ function XMLHttpResult(requette, xml, text) {
       doc_ap_ip.textContent    = ap_ip;
       doc_cli_ssid.textContent = cli_ssid;
       doc_cli_ip.textContent   = cli_ip;
+      
+      // Met à jour le réseau en cours si la page en cours est
+      // netconfig.html (présence des inputs "ssidActuel" et ipActuel"
+      var ssidActuel = document.getElementById("ssidActuel");
+      var ipActuel = document.getElementById("ipActuel");
+      if (ssidActuel != null) {
+        ssidActuel.value = cli_ssid;
+      }
+      if (ipActuel != null) {
+        ipActuel.value = cli_ip;
+      }
 
     } else if (requette == "getvalues") {
       var masse_ba     = xml.getElementsByTagName("ba")[0].childNodes[0].nodeValue;
@@ -158,6 +359,17 @@ function XMLHttpResult(requette, xml, text) {
       input_paf_ba.value  = paf_ba;
       input_entraxe.value = entraxe;
       input_tare.value    = tare;
+    } else if (requette == "getnetworks") {
+      // Rempli la liste des réseaux disponibles
+      setNetworkList(xml);
+      attenteOK = true;
+    } else if (requette == "wificonnect") {
+      result = xml.getElementsByTagName("result")[0].childNodes[0].nodeValue;
+      if (result == "OK") {
+        location.reload();
+      } else {
+        alert("Erreur de connexion : \n" + result);
+      }
     }
   }
   if (requette == "getvalues") {
@@ -228,10 +440,11 @@ function afficheDialog(dialog_id) {
 }
 
 async function etalonnage() {
-  var divWait = document.getElementById("attente0");
-  var msg = "";
+  var divWait    = document.getElementById("attente0");
+  var msg        = "";
   var chkRefresh = document.getElementById("auto_refresh");
-  
+  var tare       = document.getElementById("tare").value;
+
   // Désactive le rafraichissement automatique
   chkRefresh.checked = false;
   // Prépare la boite de dialogue
@@ -245,7 +458,7 @@ async function etalonnage() {
   document.getElementById("dlgContenu0").innerHTML = msg;
   // Bloque les mesures sur la balance
   XMLHttpRequest_get("stopmesure");
-  XMLHttpRequest_get("affichage?text=Etalonnage r\x82seau\nen cours...");
+  XMLHttpRequest_get("affichage?text=Etalonnage réseau\nen cours...");
   
   // Attente click sur bouton suite
   while (!suiteOK) {
@@ -258,7 +471,7 @@ async function etalonnage() {
   document.getElementById('btnClose').disabled = true;
   
   // Affiche la suite
-  msg = "<p>Remise à zero<br />des balances...</p>";
+  msg = "<p>Remise à zéro<br />des balances...</p>";
   document.getElementById("dlgContenu0").innerHTML = msg;
 
   // Préparation de l'étalonnage (reset scale) et remise à zéro des balances (tare)
@@ -278,9 +491,9 @@ async function etalonnage() {
   attenteOK = false;
   divWait.classList.add("noshow");
 
-  XMLHttpRequest_get("affichage?text=Etalonnage r\x82seau\nen cours...");
+  XMLHttpRequest_get("affichage?text=Etalonnage réseau\nen cours...");
 
-  msg = "<p>Placez une masse de<br />500 g sur la balance<br />coté bord d'attaque, puis<br />Cliquez sur \"Suivant\".</p>"; 
+  msg = "<p>Placez une masse de<br />" + tare + " g sur la balance<br />coté bord d'attaque, puis<br />Cliquez sur \"Suivant\".</p>"; 
   document.getElementById("dlgContenu0").innerHTML = msg;
 
   // Attente click sur bouton suite
@@ -304,9 +517,9 @@ async function etalonnage() {
   attenteOK = false;
   divWait.classList.add("noshow");
 
-  XMLHttpRequest_get("affichage?text=Etalonnage r\x82seau\nen cours...");
+  XMLHttpRequest_get("affichage?text=Etalonnage réseau\nen cours...");
 
-  msg = "<p>Placez une masse de<br />500 g sur la balance<br />coté bord de fuite, puis<br/>Cliquez sur \"Suivant\".</p>"; 
+  msg = "<p>Placez une masse de<br />" + tare + " g sur la balance<br />coté bord de fuite, puis<br/>Cliquez sur \"Suivant\".</p>"; 
   document.getElementById("dlgContenu0").innerHTML = msg;
 
   // Attente click sur bouton suite
@@ -373,9 +586,14 @@ function reboot() {
   var divWait = document.getElementById("attente0");
   dlgMask.classList.remove("noshow");
   divWait.classList.remove("noshow");
-  // Reboot
-  XMLHttpRequest_get("reboot");
-  setTimeout(function() { document.location.replace("/") }, 15000);
+  if (confirm("Redémarrer la balance ?")) {
+    // Reboot
+    XMLHttpRequest_get("reboot");
+    setTimeout(function() { document.location.replace("/") }, 15000);
+  } else {
+    dlgMask.classList.add("noshow");
+    divWait.classList.add("noshow");
+  }
   
 }
